# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Seattle Solvers Tools API",
    "description": "FTC parts inventory, search, and receipt reading for tools.seattlesolvers.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/teams": {
      "post": {
        "summary": "Create a new team and get its one-time invite code",
        "tags": [
          "auth"
        ],
        "description": "Creates only the Team — no login account, no token. Every person, including whoever calls this, gets their own account the same way everyone else does: POST /auth/join with the returned inviteCode. The first person to join a new team becomes TEAM_ADMIN automatically; everyone after that joins as MEMBER. There is no other way to see this invite code afterward (rotating it requires already being a team admin, which requires already having an account) — the caller MUST show `warning` and have the user save the code before leaving the page. Everyone after the first joiner lands as a read-only VIEWER — see POST /auth/join.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "teamNumber": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "teamName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": [
                  "teamNumber",
                  "teamName"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "anthropicApiKeyConfigured": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name",
                        "inviteCode",
                        "anthropicApiKeyConfigured"
                      ],
                      "additionalProperties": false
                    },
                    "warning": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "team",
                    "warning"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/join": {
      "post": {
        "summary": "Join an existing team using its invite code",
        "tags": [
          "auth"
        ],
        "description": "The usual way to get an account: teams have no login of their own, so this is how every member — including whoever created the team — signs up. The first person to join a brand-new team (0 existing members) becomes TEAM_ADMIN automatically; everyone after that joins as a read-only VIEWER until an admin promotes them (see PATCH /teams/members/:userId).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 40
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 200
                  }
                },
                "required": [
                  "inviteCode",
                  "displayName",
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "VIEWER",
                            "MEMBER",
                            "TEAM_ADMIN",
                            "SUPER_ADMIN"
                          ]
                        },
                        "teamId": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "teamId"
                      ],
                      "additionalProperties": false
                    },
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "anthropicApiKeyConfigured": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name",
                        "inviteCode",
                        "anthropicApiKeyConfigured"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "token",
                    "user",
                    "team"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "VIEWER",
                            "MEMBER",
                            "TEAM_ADMIN",
                            "SUPER_ADMIN"
                          ]
                        },
                        "teamId": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "teamId"
                      ],
                      "additionalProperties": false
                    },
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "anthropicApiKeyConfigured": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name",
                        "inviteCode",
                        "anthropicApiKeyConfigured"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "token",
                    "user",
                    "team"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Get the current user and their team",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "VIEWER",
                            "MEMBER",
                            "TEAM_ADMIN",
                            "SUPER_ADMIN"
                          ]
                        },
                        "teamId": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "teamId"
                      ],
                      "additionalProperties": false
                    },
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "anthropicApiKeyConfigured": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name",
                        "inviteCode",
                        "anthropicApiKeyConfigured"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "user",
                    "team"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/profile": {
      "patch": {
        "summary": "Update your own display name",
        "tags": [
          "auth"
        ],
        "description": "The name shown next to receipts you upload and next to your entry in the team member list. Changes take effect immediately; no re-login needed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80
                  }
                },
                "required": [
                  "displayName"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "VIEWER",
                            "MEMBER",
                            "TEAM_ADMIN",
                            "SUPER_ADMIN"
                          ]
                        },
                        "teamId": {
                          "type": "string",
                          "nullable": true
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "teamId"
                      ],
                      "additionalProperties": false
                    },
                    "team": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "number": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "inviteCode": {
                          "type": "string"
                        },
                        "anthropicApiKeyConfigured": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "name",
                        "inviteCode",
                        "anthropicApiKeyConfigured"
                      ],
                      "additionalProperties": false,
                      "nullable": true
                    }
                  },
                  "required": [
                    "user",
                    "team"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/invite-code": {
      "get": {
        "summary": "Get the current team's invite code",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "number": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "anthropicApiKeyConfigured": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "number",
                    "name",
                    "inviteCode",
                    "anthropicApiKeyConfigured"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/join": {
      "post": {
        "summary": "Join a team with an invite code, using your existing account",
        "tags": [
          "teams"
        ],
        "description": "For accounts that do not currently belong to a team. To sign up and join in one step, use POST /auth/join instead. Leave your current team first if you have one. The first person to join a brand-new team (0 existing members) becomes TEAM_ADMIN automatically, same as POST /auth/join.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inviteCode": {
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 40
                  }
                },
                "required": [
                  "inviteCode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "number": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "inviteCode": {
                      "type": "string"
                    },
                    "anthropicApiKeyConfigured": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "number",
                    "name",
                    "inviteCode",
                    "anthropicApiKeyConfigured"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}